requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.4
Pillow==6.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
urllib3==1.25.3
Werkzeug==0.15.4
